% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_col}
\alias{geom_col}
\title{geom_col in the Open Justice Oklahoma style}
\usage{
geom_col(
  mapping = NULL,
  data = NULL,
  position = "stack",
  ...,
  width = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
A custom version of the geom_col() geom that uses the OJO version of the ggproto object instead of the default
(this makes the bars skinnier and nice looking consistently)
}
