% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{GeomColOJO}
\alias{GeomColOJO}
\title{OJO version of the ggproto GeomCol object}
\format{
An object of class \code{GeomCol} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomColOJO
}
\description{
This is a near-exact copy of the default ggplot2 GeomCol ggproto object.
The only difference is that I've adjusted it to make the default column width smaller.
This ensures that it looks right no matter what the scale of the data you're using
}
\keyword{datasets}
