% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_gt.R
\name{gt_base}
\alias{gt_base}
\title{gt_base Function}
\usage{
gt_base(
  data,
  title = NA,
  subtitle = NA,
  font_size = 14,
  format_cols = TRUE,
  analyst_name = NA,
  source = NA
)
}
\arguments{
\item{data}{A data frame to be converted into a gt table.}

\item{title}{Optional. A title for the table. Default is NA.}

\item{subtitle}{Optional. A subtitle for the table. Default is NA.}

\item{font_size}{The font size for the table. Default is 14.}

\item{format_cols}{Should gt::fmt_auto() be applied to all cols?}

\item{analyst_name}{The name of the analyst to credit in the footnote}

\item{source}{The source / domain of the data}
}
\value{
A gt table based on the input data frame with specified modifications.
}
\description{
gt_base Function
}
\examples{
\dontrun{
gt_base(data = mtcars, title = "Motor Trend Car Road Tests", subtitle = "From mtcars")
}
}
