% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_labs.R
\name{ojo_analyst_name_text}
\alias{ojo_analyst_name_text}
\title{OJO analyst name text}
\usage{
ojo_analyst_name_text(analyst_name = NA)
}
\arguments{
\item{analyst_name}{The name of the analyst to credit}
}
\value{
A string with the analyst credit text
}
\description{
A function to construct the analyst credit text
}
