% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ojo_color_pal}
\alias{ojo_color_pal}
\title{The Open Justice Oklahoma \link{ggplot2} theme}
\usage{
ojo_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used in the Open Justice Oklahoma.
}
