% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_gt.R
\name{ojo_gt}
\alias{ojo_gt}
\title{ojo_gt Function}
\usage{
ojo_gt(data, title = NA, subtitle = NA)
}
\arguments{
\item{data}{A data frame to be converted into a gt table.}

\item{title}{Optional. A title for the table. Default is NA.}

\item{subtitle}{Optional. A subtitle for the table. Default is NA.}
}
\value{
A gt table based on the input data frame with specified modifications.
}
\description{
ojo_gt Function
}
\examples{
\dontrun{
ojo_gt(data = mtcars, title = "Motor Trend Car Road Tests", subtitle = "From mtcars")
}
}
