% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_labs.R
\name{ojo_gt_captions}
\alias{ojo_gt_captions}
\title{OJO gt caption / source note with defaults}
\usage{
ojo_gt_captions(x, analyst_name = NA, source = NA)
}
\arguments{
\item{x}{A gt object}

\item{analyst_name}{The name of the analyst to credit}

\item{source}{The data source / source of the data.}
}
\description{
Wrapper for gt::tab_source_note() and gt::tab_footnote() with consistent defaults.
}
