% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_labs.R
\name{ojo_labs}
\alias{ojo_labs}
\title{OJO ggplot labels with defaults}
\usage{
ojo_labs(..., analyst_name = NA, source = NA)
}
\arguments{
\item{analyst_name}{The name of the analyst to credit}

\item{source}{The data source / source of the data.}
}
\description{
Wrapper for ggplot2::labs() with consistent defaults.
}
