% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{ojo_make_caption}
\alias{ojo_make_caption}
\title{Helper functions with sensible defaults
OJO caption text}
\usage{
ojo_make_caption(analyst_name = NA, domain = NA)
}
\arguments{
\item{domain}{The type / domain of the data. Must be "oscn", "ocdc", "ppb",}

\item{name}{The name of the analyst to credit}
}
\description{
Creates the text for a caption to add to ggplots and gt tables, including consistent default "source: " statements.
}
