% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_main}
\alias{palette_ojo_main}
\alias{palette_ojo_diverging}
\alias{palette_ojo_quintile}
\alias{palette_ojo_politics}
\alias{palette_ojo_cyan}
\alias{palette_ojo_gray}
\alias{palette_ojo_yellow}
\alias{palette_ojo_magenta}
\alias{palette_ojo_green}
\alias{palette_ojo_spacegray}
\alias{palette_ojo_red}
\title{Urban Institute palettes}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 5.

An object of class \code{character} of length 2.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\source{
Open Justice Oklahoma Data Visualization Style Guide
\url{http://openjusticeok.github.io/styleguide/}
}
\usage{
palette_ojo_main

palette_ojo_diverging

palette_ojo_quintile

palette_ojo_politics

palette_ojo_cyan

palette_ojo_gray

palette_ojo_yellow

palette_ojo_magenta

palette_ojo_green

palette_ojo_spacegray

palette_ojo_red
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Open Justice Oklahoma Data Visualization Style Guide.
\url{http://openjusticeok.github.io/styleguide/}
}
\concept{ojo palettes}
\keyword{datasets}
