% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_set_theme.R
\name{ojo_set_theme}
\alias{ojo_set_theme}
\title{Use Oklahoma Policy Institute \link{ggplot2} themes}
\usage{
ojo_set_theme(
  theme = "okpi",
  base_size = 18,
  base_line_size = 0.5,
  base_rect_size = 0.5,
  scale = "continuous"
)
}
\arguments{
\item{theme}{The theme you wish to use. Options are "okpi", "ojo", or "tok".}

\item{base_size}{The base font size for the theme. All fonts are relative to
this value.}

\item{base_line_size}{The base line size for the theme. All line sizes are
relative to this value.}

\item{base_rect_size}{The base rect size for the theme. All rect sizes are
relative to this value.}

\item{scale}{For \code{theme_ojo_map()}. Should the legend theme be continuous or discrete?}
}
\description{
\code{ojo_set_theme} provides a \link{ggplot2} theme formatted according to the
Oklahoma Policy Institute style guide, with sensible defaults.
}
