% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_labs.R
\name{ojo_source_text}
\alias{ojo_source_text}
\title{OJO source text}
\usage{
ojo_source_text(source = NA)
}
\arguments{
\item{source}{The domain / source of the data. Can be one of "oscn", "ocdc", or "ppb" for canned text, NA for no source, or a custom string.}
}
\value{
A string with the source text
}
\description{
A function to construct the "Source:" note text
}
