% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojothemes-package.R
\docType{package}
\name{ojothemes-package}
\alias{ojothemes}
\alias{ojothemes-package}
\title{ojothemes: OJO, OKPI, and TOK themes for ggplot2 and gt}
\description{
This package contains themes for ggplot2 plots and gt tables that follow the styles of OJO / OKPI / TOK.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openjusticeok.github.io/ojothemes/}
}

}
\author{
\strong{Maintainer}: Andrew Bell \email{abell@okpolicy.com}

Authors:
\itemize{
  \item Anthony Flores \email{aflores@okpolicy.com}
  \item Brancen Gregory \email{brancengregory@gmail.com}
}

}
\keyword{internal}
