% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{okpi_blue_light}
\alias{okpi_blue_light}
\title{OKPI blue light}
\format{
An object of class \code{character} of length 1.
}
\usage{
okpi_blue_light
}
\description{
OKPI blue light
}
\keyword{datasets}
