% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_okpi_main}
\alias{palette_okpi_main}
\alias{okpi_blue_palette}
\alias{okpi_blue}
\alias{okpi_red_palette}
\alias{okpi_red}
\alias{okpi_yellow_palette}
\alias{okpi_yellow}
\title{OKPI main palette}
\format{
An object of class \code{character} of length 14.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
palette_okpi_main

okpi_blue

okpi_red

okpi_yellow
}
\description{
OKPI Pallettes ==============================================================
OKPI Extended Palette

OKPI Blue Palette

OKPI Red Palette

OKPI Yellow Palette
}
\details{
A vector with hex-color codes that correspond to the extended color palette outlined in the Open Justice Oklahoma Data Visualization Style Guide.
\url{http://openjusticeok.github.io/styleguide/}

A palette of blue shades for OKPI.

A palette of red shades for OKPI.

A palette of yellow shades for OKPI.
}
\concept{okpi palettes}
\keyword{datasets}
