% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{okpi_red_light}
\alias{okpi_red_light}
\title{OKPI Red light}
\format{
An object of class \code{character} of length 1.
}
\usage{
okpi_red_light
}
\description{
OKPI Red light
}
\keyword{datasets}
