% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{okpi_yellow_light}
\alias{okpi_yellow_light}
\title{OKPI yellow light}
\format{
An object of class \code{character} of length 1.
}
\usage{
okpi_yellow_light
}
\description{
OKPI yellow light
}
\keyword{datasets}
