% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_ojo.R
\docType{data}
\name{palette_ojo}
\alias{palette_ojo}
\title{The Open Justice Oklahoma \link{ggplot2} theme}
\format{
A \code{list}.
}
\usage{
palette_ojo
}
\description{
Palette data for the Open Justice Oklahoma
}
\details{
Data used by the palettes in the uithemes package.
}
\keyword{datasets}
