% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_diverging}
\alias{palette_ojo_diverging}
\title{OJO Diverging Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_diverging
}
\description{
OJO Diverging Palette
}
\details{
A vector with hex-color codes for the OJO diverging palette.
}
\keyword{datasets}
