% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_gray}
\alias{palette_ojo_gray}
\title{OJO Gray Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_gray
}
\description{
OJO Gray Palette
}
\details{
A vector with hex-color codes for the OJO gray palette.
}
\keyword{datasets}
