% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_green}
\alias{palette_ojo_green}
\title{OJO Green Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_green
}
\description{
OJO Green Palette
}
\details{
A vector with hex-color codes for the OJO green palette.
}
\keyword{datasets}
