% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_magenta}
\alias{palette_ojo_magenta}
\title{OJO Magenta Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_magenta
}
\description{
OJO Magenta Palette
}
\details{
A vector with hex-color codes for the OJO magenta palette.
}
\keyword{datasets}
