% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_main}
\alias{palette_ojo_main}
\title{OJO Main Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_main
}
\description{
OJO Main Palette
}
\keyword{datasets}
