% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_politics}
\alias{palette_ojo_politics}
\title{OJO Politics Palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
palette_ojo_politics
}
\description{
OJO Politics Palette
}
\details{
A vector with hex-color codes for the OJO politics palette.
}
\keyword{datasets}
