% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_quintile}
\alias{palette_ojo_quintile}
\title{OJO Quintile Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
palette_ojo_quintile
}
\description{
OJO Quintile Palette
}
\details{
A vector with hex-color codes for the OJO quintile palette.
}
\keyword{datasets}
