% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_red}
\alias{palette_ojo_red}
\title{OJO Red Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_red
}
\description{
OJO Red Palette
}
\details{
A vector with hex-color codes for the OJO red palette.
}
\keyword{datasets}
