% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_spacegray}
\alias{palette_ojo_spacegray}
\title{OJO Space Gray Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_spacegray
}
\description{
OJO Space Gray Palette
}
\details{
A vector with hex-color codes for the OJO space gray palette.
}
\keyword{datasets}
