% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_yellow}
\alias{palette_ojo_yellow}
\title{OJO Yellow Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
palette_ojo_yellow
}
\description{
OJO Yellow Palette
}
\details{
A vector with hex-color codes for the OJO yellow palette.
}
\keyword{datasets}
