% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_okpi_main}
\alias{palette_okpi_main}
\title{OKPI main palette}
\format{
An object of class \code{character} of length 14.
}
\usage{
palette_okpi_main
}
\description{
OKPI Palettes
}
\details{
A vector with hex-color codes that correspond to the extended color palette outlined in the Open Justice Oklahoma Data Visualization Style Guide.
\url{http://openjusticeok.github.io/styleguide/}
}
\keyword{datasets}
