% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_okpi}
\alias{scale_color_okpi}
\title{OKPI Color Scale}
\usage{
scale_color_okpi()
}
\value{
A ggplot2 scale object.
}
\description{
This function returns a ggplot2 color scale using the OKPI main palette.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(gear))) +
  geom_point(size = 3) +
  scale_color_okpi()
}
