% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ojo.R
\name{theme_ojo_base}
\alias{theme_ojo_base}
\title{A \link{ggplot2} theme formatted in the Open Justice Oklahoma style}
\usage{
theme_ojo_base(
  base_size = 16,
  base_family = "Roboto Mono",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_ojo} provides a \link{ggplot2} theme formatted according to the
Open Justice Oklahoma style guide for web, with sensible defaults.
}
