% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_tok_main}
\alias{palette_tok_main}
\title{TOK main palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
palette_tok_main
}
\description{
TOK main palette
}
\concept{tok palettes}
\keyword{datasets}
