% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_labs.R
\name{ojo_make_caption}
\alias{ojo_make_caption}
\title{OJO caption text}
\usage{
ojo_make_caption(analyst_name = NA, source = NA)
}
\arguments{
\item{analyst_name}{The name of the analyst to credit}

\item{source}{The domain / source of the data. Can be one of "oscn", "ocdc", or "ppb" for canned text, NA for no source, or a custom string.}
}
\value{
A string with the caption text
}
\description{
Creates the text for a caption to add to ggplots and gt tables, including consistent default "source: " statements.
}
