% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_ojo_main}
\alias{palette_ojo_main}
\alias{palette_ojo_diverging}
\alias{palette_ojo_quintile}
\alias{palette_ojo_politics}
\alias{palette_ojo_cyan}
\alias{palette_ojo_gray}
\alias{palette_ojo_yellow}
\alias{palette_ojo_magenta}
\alias{palette_ojo_green}
\alias{palette_ojo_spacegray}
\alias{palette_ojo_red}
\title{OJO Main Palette}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 5.

An object of class \code{character} of length 2.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
palette_ojo_main

palette_ojo_diverging

palette_ojo_quintile

palette_ojo_politics

palette_ojo_cyan

palette_ojo_gray

palette_ojo_yellow

palette_ojo_magenta

palette_ojo_green

palette_ojo_spacegray

palette_ojo_red
}
\description{
OJO Main Palette

OJO Diverging Palette

OJO Quintile Palette

OJO Politics Palette

OJO Cyan Palette

OJO Gray Palette

OJO Yellow Palette

OJO Magenta Palette

OJO Green Palette

OJO Space Gray Palette

OJO Red Palette
}
\details{
A vector with hex-color codes for the main OJO palette.

A vector with hex-color codes for the OJO diverging palette.

A vector with hex-color codes for the OJO quintile palette.

A vector with hex-color codes for the OJO politics palette.

A vector with hex-color codes for the OJO cyan palette.

A vector with hex-color codes for the OJO gray palette.

A vector with hex-color codes for the OJO yellow palette.

A vector with hex-color codes for the OJO magenta palette.

A vector with hex-color codes for the OJO green palette.

A vector with hex-color codes for the OJO space gray palette.

A vector with hex-color codes for the OJO red palette.
}
\concept{ojo palettes}
\keyword{datasets}
